:passport_control: feat(DeploymentEnvironmentController): Implement authorization.

+++ src/main/java/io/passport/server/controller/DeploymentEnvironmentController.java
@@ -1,14 +1,19 @@
 package io.passport.server.controller;
 
 import io.passport.server.model.DeploymentEnvironment;
+import io.passport.server.model.Role;
 import io.passport.server.service.DeploymentEnvironmentService;
+import io.passport.server.service.RoleCheckerService;
+import org.keycloak.KeycloakPrincipal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
 import java.util.Optional;
 
 /**
@@ -25,21 +30,34 @@ public class DeploymentEnvironmentController {
      */
     private final DeploymentEnvironmentService deploymentEnvironmentService;
 
+    /**
+     * Role checker service for authorization
+     */
+    private final RoleCheckerService roleCheckerService;
+
     @Autowired
-    public DeploymentEnvironmentController(DeploymentEnvironmentService deploymentEnvironmentService) {
+    public DeploymentEnvironmentController(DeploymentEnvironmentService deploymentEnvironmentService, RoleCheckerService roleCheckerService) {
         this.deploymentEnvironmentService = deploymentEnvironmentService;
+        this.roleCheckerService = roleCheckerService;
     }
-
-
 
     /**
      * Read DeploymentEnvironment by environmentId
      * @param environmentId ID of the deployment environment.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @GetMapping("/{environmentId}")
     public ResponseEntity<?> getDeploymentEnvironmentByEnvironmentId(
-            @PathVariable("environmentId") Long environmentId) {
+            @PathVariable("environmentId") Long environmentId,
+            @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
+
+        // Allowed roles for this endpoint
+        List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+        // Check role of the user
+        if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         Optional<DeploymentEnvironment> deploymentEnvironment = this.deploymentEnvironmentService.findDeploymentEnvironmentById(environmentId);
 
@@ -54,11 +72,21 @@ public class DeploymentEnvironmentController {
     /**
      * Create a Deployment environment.
      * @param deploymentEnvironment DeploymentEnvironment model instance to be created.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @PostMapping()
-    public ResponseEntity<?> createDeploymentEnvironment(@RequestBody DeploymentEnvironment deploymentEnvironment) {
+    public ResponseEntity<?> createDeploymentEnvironment(@RequestBody DeploymentEnvironment deploymentEnvironment,
+                                                         @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try{
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             DeploymentEnvironment savedDevelopmentEnvironment = this.deploymentEnvironmentService
                                                 .saveDevelopmentEnvironment(deploymentEnvironment);
             return ResponseEntity.status(HttpStatus.CREATED).body(savedDevelopmentEnvironment);
@@ -73,11 +101,22 @@ public class DeploymentEnvironmentController {
      * Update Deployment environment.
      * @param deploymentEnvironmentId ID of the deployment environment that is to be updated.
      * @param updatedDeploymentEnvironment DeploymentEnvironment model instance with updated details.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @PutMapping("/{deploymentEnvironmentId}")
-    public ResponseEntity<?> updateDeploymentEnvironment(@PathVariable Long deploymentEnvironmentId, @RequestBody DeploymentEnvironment updatedDeploymentEnvironment) {
+    public ResponseEntity<?> updateDeploymentEnvironment(@PathVariable Long deploymentEnvironmentId,
+                                                         @RequestBody DeploymentEnvironment updatedDeploymentEnvironment,
+                                                         @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try{
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             Optional<DeploymentEnvironment> savedDeploymentEnvironment = this.deploymentEnvironmentService.updateDeploymentEnvironment(deploymentEnvironmentId, updatedDeploymentEnvironment);
             if(savedDeploymentEnvironment.isPresent()) {
                 return ResponseEntity.ok().body(savedDeploymentEnvironment.get());
@@ -94,11 +133,21 @@ public class DeploymentEnvironmentController {
     /**
      * Delete a deployment environment by DeploymentEnvironment ID.
      * @param deploymentEnvironmentId ID of the deployment environment that is to be deleted.
+     * @param principal KeycloakPrincipal object that holds access token
      * @return
      */
     @DeleteMapping("/{deploymentEnvironmentId}")
-    public ResponseEntity<?> deletePersonnel(@PathVariable Long deploymentEnvironmentId) {
+    public ResponseEntity<?> deletePersonnel(@PathVariable Long deploymentEnvironmentId,
+                                             @AuthenticationPrincipal KeycloakPrincipal<?> principal) {
         try{
+
+            // Allowed roles for this endpoint
+            List<Role> allowedRoles = List.of(Role.DATA_SCIENTIST);
+            // Check role of the user
+            if(!this.roleCheckerService.hasAnyRole(principal, allowedRoles)){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+
             boolean isDeleted = this.deploymentEnvironmentService.deleteDeploymentEnvironment(deploymentEnvironmentId);
             if(isDeleted) {
                 return ResponseEntity.noContent().build();
